requirements.txt
py-solc-x
psutil
